refactor: move pages folder and files
