+++ src/pages/about/About.scss
@@ -1,0 +1,5 @@
+.intro {
+  width: 50%;
+  text-align: center;
+  margin-bottom: 1rem;
+}
