+++ src/pages/editor/Editor.scss
