+++ src/pages/home/Home.scss
@@ -1,0 +1,14 @@
+.intro {
+  width: 50%;
+  text-align: center;
+  margin-bottom: 1rem;
+}
+
+.links {
+  display: flex;
+  flex-direction: row;
+
+  a {
+    margin-right: 1rem;
+  }
+}
